Updated name from Game_Service_Impl to GameServiceImpl, removed unnecessary methods, removed "javaGameService" from annotation
